push-campaign/requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
databricks-sql-connector>=3.0.0
google-auth>=2.0.0
gspread>=6.0.0
